+++ apps/api/database/seeders/DatabaseSeeder.php
@@ -23,5 +23,7 @@ class DatabaseSeeder extends Seeder
 
             $this->call(DemoSeeder::class);
         }
+
+        $this->call(RoleSeeder::class);
     }
 }